# example/api/product/app/v1/v1.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "product/app/v1/v1.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "BlogService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/articles": {
      "get": {
        "operationId": "BlogService_GetArticles",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetArticlesResp"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "title",
            "description": "@inject_tag: form:\"title\"",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "page",
            "description": "@inject_tag: form:\"page\"",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "page_size",
            "description": "@inject_tag: form:\"page_size\"",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "author_id",
            "description": "@inject_tag: form:\"author_id\" uri:\"author_id\"",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "BlogService"
        ]
      }
    },
    "/v1/articles/{article_id}/download": {
      "get": {
        "summary": "自定义响应示例：文件下载",
        "operationId": "BlogService_DownloadArticle",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "article_id",
            "description": "@inject_tag: uri:\"article_id\"",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "BlogService"
        ]
      }
    },
    "/v1/articles/{article_id}/xml": {
      "get": {
        "summary": "自定义响应示例：XML 输出",
        "operationId": "BlogService_GetArticleXML",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "article_id",
            "description": "@inject_tag: uri:\"article_id\"",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "BlogService"
        ]
      }
    },
    "/v1/author/{author_id}/articles": {
      "get": {
        "operationId": "BlogService_GetArticles2",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetArticlesResp"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "author_id",
            "description": "@inject_tag: form:\"author_id\" uri:\"author_id\"",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "title",
            "description": "@inject_tag: form:\"title\"",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "page",
            "description": "@inject_tag: form:\"page\"",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "page_size",
            "description": "@inject_tag: form:\"page_size\"",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "BlogService"
        ]
      },
      "post": {
        "operationId": "BlogService_CreateArticle",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1Article"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "author_id",
            "description": "@inject_tag: form:\"author_id\" uri:\"author_id\"",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "title",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "content",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "BlogService"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1Article": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "content": {
          "type": "string"
        },
        "author_id": {
          "type": "integer",
          "format": "int32",
          "title": "@inject_tag: form:\"author_id\" uri:\"author_id\""
        }
      }
    },
    "v1GetArticlesResp": {
      "type": "object",
      "properties": {
        "total": {
          "type": "string",
          "format": "int64"
        },
        "articles": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Article"
          }
        }
      }
    }
  }
}