# example/api/product/app/v1/v1_gin.pb.go
// Code generated by github.com/mohuishou/protoc-gen-go-gin. DO NOT EDIT.

package v1

import (
	errors "errors"
	gin "github.com/gin-gonic/gin"
	protojson "google.golang.org/protobuf/encoding/protojson"
	emptypb "google.golang.org/protobuf/types/known/emptypb"
	io "io"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the mohuishou/protoc-gen-go-gin package it is being compiled against.
//gin.errors.
// protojson.emptypb.io.

var _ = io.EOF // 防止 "imported and not used" 报错

var marshaler = protojson.MarshalOptions{
	UseProtoNames:   true,
	EmitUnpopulated: true,
}
var unmarshaler = protojson.UnmarshalOptions{
	AllowPartial:   false,
	DiscardUnknown: true,
}

type BlogServiceHTTPServer interface {
	CreateArticle(*gin.Context, *Article) (*Article, error)

	DownloadArticle(*gin.Context, *DownloadArticleReq) (*emptypb.Empty, error)

	GetArticleXML(*gin.Context, *GetArticleXMLReq) (*emptypb.Empty, error)

	GetArticles(*gin.Context, *GetArticlesReq) (*GetArticlesResp, error)
}

func RegisterBlogServiceHTTPServer(r gin.IRouter, srv BlogServiceHTTPServer) {
	s := BlogService{
		server: srv,
		router: r,
		resp:   defaultBlogServiceResp{},
	}
	s.RegisterService()
}

type BlogService struct {
	server BlogServiceHTTPServer
	router gin.IRouter
	resp   interface {
		Error(ctx *gin.Context, err error)
		ParamsError(ctx *gin.Context, err error)
		Success(ctx *gin.Context, data interface{})
	}
}

// Resp 返回值
type defaultBlogServiceResp struct{}

func (resp defaultBlogServiceResp) response(ctx *gin.Context, status, code int, msg string, data interface{}) {
	ctx.JSON(status, map[string]interface{}{
		"code": code,
		"msg":  msg,
		"data": data,
	})
}

// Error 返回错误信息
func (resp defaultBlogServiceResp) Error(ctx *gin.Context, err error) {
	code := -1
	status := 500
	msg := "Unknown error"

	if err == nil {
		msg += ", err is nil"
		resp.response(ctx, status, code, msg, nil)
		return
	}

	type iCode interface {
		HTTPCode() int
		Message() string
		Code() int
	}

	var c iCode
	if errors.As(err, &c) {
		status = c.HTTPCode()
		code = c.Code()
		msg = c.Message()
	}

	type iError interface {
		GetCode() int32
		GetMessage() string
	}

	var e iError
	if errors.As(err, &e) {
		status = 200
		code = int(e.GetCode())
		msg = e.GetMessage()
	}

	_ = ctx.Error(err)

	resp.response(ctx, status, code, msg, nil)
}

// ParamsError 参数错误
func (resp defaultBlogServiceResp) ParamsError(ctx *gin.Context, err error) {
	_ = ctx.Error(err)
	resp.response(ctx, 400, 400, "Parameter error", nil)
}

// Success 返回成功信息
func (resp defaultBlogServiceResp) Success(ctx *gin.Context, data interface{}) {
	resp.response(ctx, 200, 0, "Success", data)
}

func (s *BlogService) GetArticles_0(ctx *gin.Context) {
	var in GetArticlesReq

	if err := ctx.ShouldBindUri(&in); err != nil {
		s.resp.ParamsError(ctx, err)
		return
	}

	if err := ctx.ShouldBindQuery(&in); err != nil {
		s.resp.ParamsError(ctx, err)
		return
	}

	//md := metadata.New(nil)
	//for k, v := range ctx.Request.Header {
	//	md.Set(k, v...)
	//}
	//newCtx := metadata.NewIncomingContext(ctx, md)
	out, err := s.server.(BlogServiceHTTPServer).GetArticles(ctx, &in)
	if err != nil {
		s.resp.Error(ctx, err)
		return
	}

	// 如果有返回值，输出 JSON
	if out != nil {
		s.resp.Success(ctx, out)
		return
	}

	// 如果返回 nil，检查是否已经写入响应（如文件下载、代理转发等）
	if !ctx.Writer.Written() {
		// 严格模式：开发者忘记处理响应，返回错误
		s.resp.Error(ctx, errors.New("handler returned nil but no response was written"))
	}
}

func (s *BlogService) GetArticles_1(ctx *gin.Context) {
	var in GetArticlesReq

	if err := ctx.ShouldBindQuery(&in); err != nil {
		s.resp.ParamsError(ctx, err)
		return
	}

	//md := metadata.New(nil)
	//for k, v := range ctx.Request.Header {
	//	md.Set(k, v...)
	//}
	//newCtx := metadata.NewIncomingContext(ctx, md)
	out, err := s.server.(BlogServiceHTTPServer).GetArticles(ctx, &in)
	if err != nil {
		s.resp.Error(ctx, err)
		return
	}

	// 如果有返回值，输出 JSON
	if out != nil {
		s.resp.Success(ctx, out)
		return
	}

	// 如果返回 nil，检查是否已经写入响应（如文件下载、代理转发等）
	if !ctx.Writer.Written() {
		// 严格模式：开发者忘记处理响应，返回错误
		s.resp.Error(ctx, errors.New("handler returned nil but no response was written"))
	}
}

func (s *BlogService) CreateArticle_0(ctx *gin.Context) {
	var in Article

	if err := ctx.ShouldBindUri(&in); err != nil {
		s.resp.ParamsError(ctx, err)
		return
	}

	reqRaw, err := io.ReadAll(ctx.Request.Body)
	if err != nil {
		s.resp.ParamsError(ctx, err)
		return
	}
	if err := unmarshaler.Unmarshal(reqRaw, &in); err != nil {
		s.resp.ParamsError(ctx, err)
		return
	}

	//md := metadata.New(nil)
	//for k, v := range ctx.Request.Header {
	//	md.Set(k, v...)
	//}
	//newCtx := metadata.NewIncomingContext(ctx, md)
	out, err := s.server.(BlogServiceHTTPServer).CreateArticle(ctx, &in)
	if err != nil {
		s.resp.Error(ctx, err)
		return
	}

	// 如果有返回值，输出 JSON
	if out != nil {
		s.resp.Success(ctx, out)
		return
	}

	// 如果返回 nil，检查是否已经写入响应（如文件下载、代理转发等）
	if !ctx.Writer.Written() {
		// 严格模式：开发者忘记处理响应，返回错误
		s.resp.Error(ctx, errors.New("handler returned nil but no response was written"))
	}
}

func (s *BlogService) DownloadArticle_0(ctx *gin.Context) {
	var in DownloadArticleReq

	if err := ctx.ShouldBindUri(&in); err != nil {
		s.resp.ParamsError(ctx, err)
		return
	}

	if err := ctx.ShouldBindQuery(&in); err != nil {
		s.resp.ParamsError(ctx, err)
		return
	}

	//md := metadata.New(nil)
	//for k, v := range ctx.Request.Header {
	//	md.Set(k, v...)
	//}
	//newCtx := metadata.NewIncomingContext(ctx, md)
	out, err := s.server.(BlogServiceHTTPServer).DownloadArticle(ctx, &in)
	if err != nil {
		s.resp.Error(ctx, err)
		return
	}

	// 如果有返回值，输出 JSON
	if out != nil {
		s.resp.Success(ctx, out)
		return
	}

	// 如果返回 nil，检查是否已经写入响应（如文件下载、代理转发等）
	if !ctx.Writer.Written() {
		// 严格模式：开发者忘记处理响应，返回错误
		s.resp.Error(ctx, errors.New("handler returned nil but no response was written"))
	}
}

func (s *BlogService) GetArticleXML_0(ctx *gin.Context) {
	var in GetArticleXMLReq

	if err := ctx.ShouldBindUri(&in); err != nil {
		s.resp.ParamsError(ctx, err)
		return
	}

	if err := ctx.ShouldBindQuery(&in); err != nil {
		s.resp.ParamsError(ctx, err)
		return
	}

	//md := metadata.New(nil)
	//for k, v := range ctx.Request.Header {
	//	md.Set(k, v...)
	//}
	//newCtx := metadata.NewIncomingContext(ctx, md)
	out, err := s.server.(BlogServiceHTTPServer).GetArticleXML(ctx, &in)
	if err != nil {
		s.resp.Error(ctx, err)
		return
	}

	// 如果有返回值，输出 JSON
	if out != nil {
		s.resp.Success(ctx, out)
		return
	}

	// 如果返回 nil，检查是否已经写入响应（如文件下载、代理转发等）
	if !ctx.Writer.Written() {
		// 严格模式：开发者忘记处理响应，返回错误
		s.resp.Error(ctx, errors.New("handler returned nil but no response was written"))
	}
}

func (s *BlogService) RegisterService() {

	s.router.Handle("GET", "/v1/author/:author_id/articles", s.GetArticles_0)

	s.router.Handle("GET", "/v1/articles", s.GetArticles_1)

	s.router.Handle("POST", "/v1/author/:author_id/articles", s.CreateArticle_0)

	s.router.Handle("GET", "/v1/articles/:article_id/download", s.DownloadArticle_0)

	s.router.Handle("GET", "/v1/articles/:article_id/xml", s.GetArticleXML_0)

}